Added autosuggest functionality
Added autosuggest functionality, which allows a user to type in an address, and receive suggestions based on what they've typed.

Also added is the limiter library to allow the web server to limit the amount of requests a user can perform per second.

More info of the library here: https://pypi.org/project/Flask-Limiter/