Merge branch 'shipments' of https://github.com/nestcx/freight-carbon-emission-calculator into add-shipment-endpoint-with-wtforms
Errors:
Type Error occuring when Plan shipment page is accessed.

Extra Files:
afcc/shipment.py is extra, just kept there for reference, could be deleted as soon as the error is solved and the endpoint starts working fine.

Major Changes:
Added afcc/shipment/shipmentform.html
Added afcc/shipment/forms.py
Edited the code to incorporate the wtform for shipment and for add shipment to upload everything to database.

Extra Comments:
I have not added the in-line comments to my code yet as this code still has some errors that needs to be resolved.
I tried to solve it for a long time but couldn't do it.